Fix target metadata
